+++ packages/docs/src/data/contentExplorerItems.ts
@@ -168,13 +168,13 @@ export const contentExplorerItems: TreeviewItem[] = [
 						]
 					},
 					{
-						name: 'Links',
+						name: 'LinksList',
 						children: [
 							{
 								name: 'tests',
 								children: [
 									{
-										name: 'Links.spec.ts',
+										name: 'LinksList.spec.ts',
 										file: 'ts'
 									}
 								]
@@ -184,7 +184,7 @@ export const contentExplorerItems: TreeviewItem[] = [
 								file: 'ts'
 							},
 							{
-								name: 'Links.vue',
+								name: 'LinksList.vue',
 								file: 'vue'
 							}
 						]