+++ packages/vue-cli-plugin-vue-dash/generator/template/src/components/LinksList/index.ts
@@ -1,0 +1,3 @@
+import LinksList from './LinksList.vue';
+
+export default LinksList;